pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["evergreen_mcp*", "evergreen_waterfall_triage*"]
namespaces = true

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["evergreen_mcp"]

[project]
name = "evergreen-mcp-server"
version = "0.4.2"
description = "A server for interacting with the Evergreen API."
authors = [
    { name = "Geoff Mishkin", email = "geoff.mishkin@mongodb.com" }
]
dependencies = [
    "urllib3",
    "python-dateutil",
    "aiohttp",
    "aiohttp-retry",
    "pydantic",
    "typing-extensions",
    "pyyaml",
    "mcp",
    "fastmcp>=3.0.0b1",
    "gql[aiohttp]",
    "httpx>=0.24.0",
    "pyjwt>=2.0.0",
    "cryptography>=41.0.0",
    "authlib>=1.0.0",
    "sentry-sdk>=2.43.0",
    "filelock>=3.0.0,<4"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[project.scripts]
evergreen-mcp-server = "evergreen_mcp.server:main"
evergreen-waterfall-triage = "evergreen_waterfall_triage.cli:main"
evergreen-waterfall-investigate = "evergreen_waterfall_triage.investigation_cli:main"
evergreen-restart-tasks = "evergreen_waterfall_triage.restart_cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests (require valid credentials)",
]
